pyproject.toml
[tool.poetry]
name = "csfdock"
version = "0.1.2.2"
description = "CsfDock"
authors = ["Rahul Brahma <mrbrahma.rahul@gmail.com>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/takshan/csfdock"
repository = "https://github.com/takshan/csfdock"
documentation = "https://csfdock.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.urls]
Changelog = "https://github.com/takshan/csfdock/releases"

[tool.poetry.dependencies]
python = "^3.6.1"
click = "^8.0.1"
ipykernel = "6.0.3"
ipywidgets = "7.6.4"
isort = "5.9.3"
matplotlib = "3.4.3"
numpy = "1.21.1"
pandas = "1.3.0"
pickle5 = "0.0.11"
plotly = "5.1.0"
pytz = "2021.1"
rdkit-pypi = "2021.3.4"
rich = "10.7.0"
scikit-learn = "0.24.2"
xgboost = "1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = {extras = ["toml"], version = "^5.4"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.12.1"
xdoctest = {extras = ["colors"], version = "^0.15.6"}
sphinx = "^4.1.2"
sphinx-autobuild = "^2021.3.14"
pre-commit = "^2.15.0"
flake8 = "^3.9.2"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.1"
darglint = "^1.8.0"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"
sphinx-rtd-theme = "^0.5.2"
sphinx-click = "^3.0.1"
Pygments = "^2.9.0"




[tool.poetry.scripts]
csfdock = "csfdock.ar2a_v3:main"

[tool.coverage.paths]
                        
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["csfdock"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
